+++ frontend/src/app/users/list.component.ts
@@ -2,10 +2,11 @@
 import { first } from 'rxjs/operators';
 
 import { UserService } from '@app/services';
+import {User} from "@app/models";
 
 @Component({ templateUrl: 'list.component.html' })
 export class ListComponent implements OnInit {
-    users = null;
+    public users: User[] = [];
 
     constructor(private userService: UserService) {}
 
@@ -17,6 +18,9 @@ export class ListComponent implements OnInit {
 
     deleteUser(id: number) {
         const user = this.users.find(x => x.id === id);
+        if (user === undefined) {
+            return;
+        }
         user.isDeleting = true;
         this.userService.delete(id)
             .pipe(first())
